refactor: move integration/pkg/backend package to integration/util/backend
