+++ integration/service/oasis/osrmconnector/osrmhttpclient.go
@@ -8,7 +8,7 @@ import (
 	"github.com/Telenav/osrm-backend/integration/api"
 	"github.com/Telenav/osrm-backend/integration/api/osrm/route"
 	"github.com/Telenav/osrm-backend/integration/api/osrm/table"
-	"github.com/Telenav/osrm-backend/integration/pkg/backend"
+	"github.com/Telenav/osrm-backend/integration/util/backend"
 	"github.com/golang/glog"
 )
