+++ integration/service/oasis/searchconnector/searchhttpclient.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"github.com/Telenav/osrm-backend/integration/api/search/nearbychargestation"
-	"github.com/Telenav/osrm-backend/integration/pkg/backend"
+	"github.com/Telenav/osrm-backend/integration/util/backend"
 	"github.com/golang/glog"
 )
